requirements.txt
# requirements.txt
selenium==4.18.1
undetected-chromedriver==3.5.4
webdriver-manager==4.0.1
fake-useragent==1.5.0
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
aiohttp==3.9.3
aiohttp-socks==0.8.4
colorama==0.4.6
python-dotenv==1.0.1
pyvirtualdisplay==3.0
xvfbwrapper==0.2.9
cloudscraper==1.2.71
pillow==10.2.0
numpy==1.26.4
pytz==2024.1
redis==5.0.3
schedule==1.2.0
tenacity==8.2.3
retry==0.9.2
ratelimit==2.2.1
fake-headers==1.0.2
selenium-stealth==1.0.6
2captcha-python==1.1.0  # Opzionale per captcha
discord-webhook==1.3.0  # Notifiche
telegram-send==0.34      # Notifiche
